feat(ci/push-docker): allow to automatically push the Docker image on commit (#114)
We pushed to `pan93412/unblock-netease-music-enhanced`.

It will build and push as  `develop` when received the new commit in `enhanced`; Similarly, it will build and push as `develop`, `released` and the current tag name when received the `published` event. You can also trigger it manually.

It currently builds four variant of Docker images: `linux/amd64`, `linux/arm64`, `linux/arm/v6`, and `linux/arm/v7`.

This resolved #110.